# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "get": {
        "description": "Obtiene una lista de reportes con filtrado opcional por estado y paginación.",
        "operationId": "ReportsController_getReports",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "examples": {
              "pendiente": {
                "value": "1"
              },
              "aprobada": {
                "value": "2"
              },
              "rechazada": {
                "value": "3"
              }
            },
            "schema": {
              "enum": [
                "1",
                "2",
                "3"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "examples": {
              "valid": {
                "value": "123"
              },
              "invalid": {
                "value": "abc"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "examples": {
              "valid": {
                "value": 1
              },
              "invalid": {
                "value": -1
              }
            },
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports retrieved successfully.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "title": "Reporte de sitio fraudulento",
                    "image": "report-pictures/1d92a0a7cbb8f6a32b6ff1a98ecf2af4f13293be20e38e7db23ccfca9a412b8a.jpg",
                    "description": "Este sitio web solicita datos bancarios sin medidas de seguridad y redirige a páginas falsas.",
                    "created_at": "2025-09-27T23:50:39.000Z",
                    "updated_at": "2025-09-27T23:50:39.000Z",
                    "user_name": "Skibidi Toilet",
                    "created_by": 1,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://banco-seguro-falso.com",
                    "categories": [
                      1,
                      5
                    ]
                  },
                  {
                    "id": 2,
                    "title": "Página de phishing detectada",
                    "image": "report-pictures/34baf3de9a5ef873b432bd723d9e0d45a4c2b79f915a86d2a7b5419ccf78d66c.jpg",
                    "description": "El portal imita la interfaz de una empresa de envíos para robar credenciales de acceso.",
                    "created_at": "2025-09-27T23:51:32.000Z",
                    "updated_at": "2025-09-27T23:51:32.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 2,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://envios-gratis-seguro.net",
                    "categories": [
                      2,
                      5
                    ]
                  },
                  {
                    "id": 3,
                    "title": "Sitio de ventas falsas",
                    "image": "report-pictures/57c0a3a48e923d8bb9e94b3a8ff743a58e9c4e71d4ccf0e6e2e3d513a7f49fdd.jpg",
                    "description": "La página ofrece productos electrónicos a precios demasiado bajos y nunca realiza las entregas.",
                    "created_at": "2025-09-27T23:51:51.000Z",
                    "updated_at": "2025-09-27T23:51:51.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 3,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://ofertas-electronica-barata.org",
                    "categories": [
                      3,
                      5
                    ]
                  },
                  {
                    "id": 5,
                    "title": "Plataforma de inversión fraudulenta",
                    "image": "report-pictures/b9237d0f30ecb0a7dc2f84e6a1cc3af324ea742b93a56f5dd8f0cb11e629bf5e.jpg",
                    "description": "El sitio promete ganancias irreales en criptomonedas y exige depósitos sin garantía alguna.",
                    "created_at": "2025-09-27T23:52:29.000Z",
                    "updated_at": "2025-09-27T23:52:29.000Z",
                    "user_name": "Dr. Sahur",
                    "created_by": 2,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://crypto-ganancias-rapidas.info",
                    "categories": [
                      4,
                      5
                    ]
                  },
                  {
                    "id": 7,
                    "title": "Portal falso de soporte técnico",
                    "image": "report-pictures/ccf0a7bb2e9d9a4b8347ff291c4b731ddf28573c4a2a5b5d67e3bc111f0a89f3.jpg",
                    "description": "La web se hace pasar por un servicio de soporte oficial y pide pagos por reparaciones inexistentes.",
                    "created_at": "2025-09-27T23:52:50.000Z",
                    "updated_at": "2025-09-27T23:52:50.000Z",
                    "user_name": "Skibidi Toilet",
                    "created_by": 1,
                    "user_image": "profile-pictures/default.jpg",
                    "report_url": "https://soporte-oficial-falso.com",
                    "categories": [
                      1,
                      4
                    ]
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "invalidStatus": {
                    "value": {
                      "message": "Status must be one of the following values: pendiente, aprobada, rechazada",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "invalidId": {
                    "value": {
                      "message": "ID must be a positive integer",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Report with specified ID not found.",
            "content": {
              "application/json": {
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Obtiene los reportes",
        "tags": [
          "Reports"
        ]
      },
      "post": {
        "operationId": "ReportsController_createReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Create Report",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "reportId": 48,
                    "report_category": [
                      2,
                      3
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "examples": {
                  "missingFields": {
                    "summary": "Missing required fields",
                    "value": {
                      "statusCode": 400,
                      "message": "field should not be empty",
                      "error": "Bad Request"
                    }
                  },
                  "invalidCategory": {
                    "summary": "Invalid category",
                    "value": {
                      "statusCode": 400,
                      "message": "One or more categories are invalid",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Crea un nuevo reporte",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/{id}": {
      "put": {
        "operationId": "ReportsController_updateReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Se actualizo el reporte correctamente",
            "content": {
              "application/json": {
                "example": {
                  "report": {
                    "id": 53,
                    "title": "Reporte de pagina fraudulenta",
                    "image": "/images/reports/potholes.png",
                    "description": "Solicitan informacion personal",
                    "created_at": "2025-10-24T02:33:21.000Z",
                    "updated_at": "2025-10-24T02:33:21.000Z",
                    "created_by": 1,
                    "status_id": 1,
                    "report_url": "https://example.com/report/123",
                    "is_anonymous": 0
                  },
                  "categories": [
                    2,
                    3
                  ]
                }
              }
            }
          },
          "400": {
            "description": "No hay campos para actualizar",
            "content": {
              "application/json": {
                "example": {
                  "message": "At least one field must be provided for update",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          }
        },
        "summary": "Actualiza un reporte existente",
        "tags": [
          "Reports"
        ]
      },
      "delete": {
        "operationId": "ReportsController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "404": {
            "description": "Report not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User with id 442 not found",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Elimina un reporte por su ID",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/count": {
      "get": {
        "operationId": "ReportsController_countReports",
        "parameters": [
          {
            "name": "status_id",
            "required": false,
            "in": "query",
            "description": "Filter reports by status",
            "examples": {
              "pendiente": {
                "value": "1"
              },
              "aprobada": {
                "value": "2"
              },
              "rechazada": {
                "value": "3"
              }
            },
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "2",
                "3"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Count of reports",
            "content": {
              "application/json": {
                "example": {
                  "count": 42
                }
              }
            }
          }
        },
        "summary": "Cuenta los reportes, opcionalmente por estado",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/upvotes": {
      "post": {
        "operationId": "UpvotesController_postUpvote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reportId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Like creado exitosamente"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea un like a un reporte",
        "tags": [
          "Upvotes"
        ]
      },
      "delete": {
        "operationId": "UpvotesController_deleteUpvote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reportId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Like eliminado exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          }
        },
        "summary": "Elimina un like de un reporte",
        "tags": [
          "Upvotes"
        ]
      }
    },
    "/upvotes/total": {
      "post": {
        "operationId": "UpvotesController_totalUpvotes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "postId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Total de likes obtenido exitosamente",
            "content": {
              "application/json": {
                "example": 42
              }
            }
          }
        },
        "summary": "Obtiene el total de likes de un post",
        "tags": [
          "Upvotes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Endpoints de CRUD Usuarios",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PostReportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Reporte de pagina fraudulenta"
          },
          "image": {
            "type": "string",
            "example": "/images/reports/potholes.png"
          },
          "description": {
            "type": "string",
            "example": "Solicitan informacion personal"
          },
          "created_by": {
            "type": "number",
            "example": 1,
            "description": "User ID of the creator"
          },
          "status_id": {
            "type": "number",
            "example": 1,
            "description": "Status ID of the report"
          },
          "category": {
            "example": [
              1,
              2
            ],
            "description": "Array of category IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "report_url": {
            "type": "string",
            "example": "https://example.com/report/123"
          },
          "is_anonymous": {
            "type": "number",
            "enum": [
              0,
              1
            ],
            "examples": [
              0,
              1
            ]
          }
        },
        "required": [
          "title",
          "image",
          "description",
          "status_id",
          "category",
          "report_url",
          "is_anonymous"
        ]
      },
...
      "UpdateReportDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Report Title",
            "description": "Title of the report"
          },
          "description": {
            "type": "string",
            "example": "The report is about <topic>",
            "description": "Description of the report"
          },
          "status_id": {
            "type": "number",
            "example": 2,
            "description": "Status ID of the report"
          },
          "category": {
            "example": [
              1,
              2
            ],
            "description": "Array of category IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "example": "/report-pictures/updated_image.png",
            "description": "Path to the report image"
          },
          "report_url": {
            "type": "string",
            "example": "https://example.com/report/1",
            "description": "URL of the report"
          }
        }
      },
...
    }
  }
}